Create LFPG.txt
This is Paris Airspace including LFPG and LFPO. I'm thinking about adding LFPB at a later stage maybe. I'm planning to add a readme later which will be more like a guide how I play the sector, but this will take a little more time to do, so I'd open a separate request for that. MVA I wasn't able to define yet due to limitations with defining areas and some missing coordinates. Finding a workaround there may be possible in a future update.