Raphaell: EDDH v1
I need to implement more things such as the BOGMU1N or the NOLGO4W arrival but for them i have to find out first how to do a second arrival route for the same config and Departure routes for EDHL. But for now we have EDDH v1 with 2 Airports and most of the things done for now. There will be updates in future to improve, fix and complete this file. I hope you have fun and i will look foward to your responds down here!

THIS FILE IS NOT THE FINAL

Information :
There is a Virtual Airline called GENSHIN or GNH. If you don't like it or just want to make it more real you can just remove it.